Minor change to improve the code
