+++ src/main/java/seedu/duke/Duke.java
@@ -1,20 +1,22 @@
 package seedu.duke;
 
 import seedu.duke.exception.ProcessInputException;
-import seedu.duke.shooter.MediumSkill;
+import seedu.duke.player.BeginnerSkill;
+import seedu.duke.player.MediumSkill;
 import seedu.duke.ui.Ui;
 
 public class Duke {
 
-    public Duke() {
-    }
+    public Duke() {}
 
     /**
      * Main entry-point for the java.duke.Duke application.
      */
     public static void main(String[] args) {
         Formatter.printWelcomeMsg();
-        PlayerList.L1.add(new MediumSkill("Bruno"));
+//        PlayerList.L1.add(new MediumSkill("Bruno"));
+
+        //Bruno is a player that user created
         while (Ui.getIsRunning()) {
 //            Formatter.printGoalBeforeShot(Ui.roundCount);
             try {
