+++ src/main/java/seedu/duke/PlayerList.java
@@ -1,9 +1,8 @@
 package seedu.duke;
 
-import seedu.duke.shooter.Shooter;
-
+import seedu.duke.player.Player;
 import java.util.ArrayList;
 
 public class PlayerList {
-    public static ArrayList<Shooter> L1= new ArrayList<>();
+    public static ArrayList<Player> L1= new ArrayList<>();
 }
