+++ src/main/java/seedu/duke/player/ExpertSkill.java
@@ -1,14 +1,14 @@
-package seedu.duke.shooter;
+package seedu.duke.player;
 
 import seedu.duke.Formatter;
 import seedu.duke.ui.Ui;
 
-public class MediumSkill extends Shooter {
+public class ExpertSkill extends Player {
     private String name;
     private int power = 1;
-    private final int skill = 2;
+    private final int skill = 3;
 
-    public MediumSkill(String name) {
+    public ExpertSkill(String name) {
         super(name);
     }
 
@@ -22,9 +22,9 @@ public class MediumSkill extends Shooter {
         System.out.print("Power:");
         for (int i=0; i<3;i++){
             if (i<power){
-                System.out.print(" 000");
+                System.out.print(" ###");
             }else{
-                System.out.print(" OOO");
+                System.out.print("    ");
             }
         }
         if (power==1){
@@ -41,17 +41,17 @@ public class MediumSkill extends Shooter {
         System.out.print("Skill:");
         for (int i=0; i<3;i++){
             if (i<skill){
-                System.out.print(" 000");
+                System.out.print(" ###");
             }else{
-                System.out.print(" OOO");
+                System.out.print("    ");
             }
         }
-        System.out.println(" Level-Medium");
+        System.out.println(" Level-Expert");
     }
 
     @Override
     public void printGoalBeforeShoot() {
-        Formatter.printGoalBeforeShotforMedium(Ui.roundCount);
+        Formatter.printGoalBeforeShotforExpert(Ui.roundCount);
     }
 
     @Override
